avoid unnecessary buffer allocation in `PeekReader`
The old code was allocating a buffer on every read just to immediately
copy it into the internal buffer. Now we read directly into the internal
buffer.

This should be faster since we are not allocating memory and copying
data around for no reason.

Added a basic test in addition to make sure I didn't break anything.

Signed-off-by: Onur Özkan <work@onurozkan.dev>