Make generated order deterministic and add snapshot testing
This changes the hashmaps to indexmaps that preserve order of
insertion when iterating over them. This makes the generated code
order deterministic and consistent with the definition file.

This also adds snapshot testing to ensure that the generated code
remains consistent across changes or makes it very obvious when it does
change.